+++ DiaBLE Watch/Views/Details.swift
@@ -9,7 +9,7 @@ struct Details: View, LoggingView {
 
     @State private var showingCalibrationInfoForm = false
 
-    @State private var readingCountdown: Int = 0
+    @State private var readingCountdown: Int64 = 0
     @State private var secondsSinceLastConnection: Int = 0
     @State private var minutesSinceLastReading: Int = 0
 
@@ -415,7 +415,7 @@ struct Details: View, LoggingView {
                             if app.lastConnectionDate == Date.distantPast {
                                 readingCountdown = 0
                             } else {
-                                readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                                readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                             }
                         }
                     }
