+++ DiaBLE Watch/Views/Monitor.swift
@@ -10,9 +10,9 @@ struct Monitor: View, LoggingView {
 
     @Environment(\.dismiss) var dismiss
 
-    @State private var readingCountdown: Int = 0
+    @State private var readingCountdown: Int64 = 0
     @State private var minutesSinceLastReading: Int = 0
-    @State private var onlineCountdown: Int = 0
+    @State private var onlineCountdown: Int64 = 0
 
     @State private var timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
     @State private var minuteTimer = Timer.publish(every: 60, on: .main, in: .common).autoconnect()
@@ -112,7 +112,7 @@ struct Monitor: View, LoggingView {
                                 if app.lastConnectionDate == Date.distantPast {
                                     readingCountdown = 0
                                 } else {
-                                    readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                                    readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                                 }
                             }
                         }
@@ -125,7 +125,7 @@ struct Monitor: View, LoggingView {
                                 if settings.lastOnlineDate == Date.distantPast {
                                     onlineCountdown = 0
                                 } else {
-                                    onlineCountdown = settings.onlineInterval * 60 - Int(Date().timeIntervalSince(settings.lastOnlineDate))
+                                    onlineCountdown = Int64(settings.onlineInterval * 60) - Int64(Date().timeIntervalSince(settings.lastOnlineDate))
                                 }
                             }
