+++ DiaBLE Watch/Views/OnlineView.swift
@@ -19,8 +19,8 @@ struct OnlineView: View, LoggingView {
     @Environment(History.self) var history: History
     @Environment(Settings.self) var settings: Settings
 
-    @State private var onlineCountdown: Int = 0
-    @State private var readingCountdown: Int = 0
+    @State private var onlineCountdown: Int64 = 0
+    @State private var readingCountdown: Int64 = 0
 
     @State private var showingCredentials: Bool = false
 
@@ -78,7 +78,7 @@ struct OnlineView: View, LoggingView {
                                 if settings.lastOnlineDate == Date.distantPast {
                                     onlineCountdown = 0
                                 } else {
-                                    onlineCountdown = settings.onlineInterval * 60 - Int(Date().timeIntervalSince(settings.lastOnlineDate))
+                                    onlineCountdown = Int64(settings.onlineInterval * 60) - Int64(Date().timeIntervalSince(settings.lastOnlineDate))
                                 }
                             }
                     }
@@ -102,7 +102,7 @@ struct OnlineView: View, LoggingView {
                         if app.lastConnectionDate == Date.distantPast {
                             readingCountdown = 0
                         } else {
-                            readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                            readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                         }
                     }
                 }
