+++ DiaBLE/Views/Monitor.swift
@@ -10,9 +10,9 @@ struct Monitor: View, LoggingView {
 
     @State private var showingNFCAlert = false
 
-    @State private var readingCountdown: Int = 0
+    @State private var readingCountdown: Int64 = 0
     @State private var minutesSinceLastReading: Int = 0
-    @State private var onlineCountdown: Int = 0
+    @State private var onlineCountdown: Int64 = 0
 
     @State private var timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
     @State private var minuteTimer = Timer.publish(every: 60, on: .main, in: .common).autoconnect()
@@ -100,7 +100,7 @@ struct Monitor: View, LoggingView {
                                 .font(.callout.monospacedDigit())
                                 .foregroundStyle(.orange)
                                 .onReceive(timer) { _ in
-                                    readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                                    readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                                 }
                             }
                             Text(onlineCountdown > 0 ? "\(onlineCountdown) s" : " ")
@@ -108,7 +108,7 @@ struct Monitor: View, LoggingView {
                                 .font(.callout.monospacedDigit())
                                 .foregroundStyle(.cyan)
                                 .onReceive(timer) { _ in
-                                    onlineCountdown = settings.onlineInterval * 60 - Int(Date().timeIntervalSince(settings.lastOnlineDate))
+                                    onlineCountdown = Int64(settings.onlineInterval * 60) - Int64(Date().timeIntervalSince(settings.lastOnlineDate))
                                 }
                         }
                     }
